added methods to setup core location and map view properties
