+++ MoreWithMaps/ViewController.swift
@@ -24,7 +24,6 @@ class ViewController: UIViewController {
     @IBOutlet weak var myMapView: MKMapView!
     @IBOutlet weak var myTextField: UITextField!
     
-    
     let locationManager = CLLocationManager()
     var currentCoord: CLLocationCoordinate2D?
     var steps = [MKRouteStep]()
@@ -47,26 +46,15 @@ class ViewController: UIViewController {
     
     // @IBOutlet weak var categoriesButton: UIButton!
     @IBOutlet weak var categoryBackground: UIView!
-    // Karen
-    let karen = "com.apple.ttsbundle.Karen-compact"
-    // en-AU
+
     // Daniel
     let daniel = "com.apple.ttsbundle.Daniel-compact"
     // en-GB
-    // Moira
-    let moira = "com.apple.ttsbundle.Moira-compact"
-    // en-IE
-    // Samantha
-    let samantha = "com.apple.ttsbundle.Samantha-compact"
-    // en-US
-    // Tessa
-    let tessa = "com.apple.ttsbundle.Tessa-compact"
-    // en-ZA
-    
+
     let speechSynthesizer = AVSpeechSynthesizer()
+    // Var for giving directions
     var uniqueVoice = AVSpeechSynthesisVoice()
     
-    // Var for giving directions
     @IBOutlet weak var directionsGoButton: UIButton!
     var destinationMapItem = MKMapItem()
     @IBOutlet weak var directionsLabel: UILabel!
@@ -88,31 +76,38 @@ class ViewController: UIViewController {
         setUpDisplay()
         ref = Database.database().reference()
         getInitialData()
+        setupLocationManager()
+        setupMapViewProperties()
         
-        // LOCATION MANAGER SETUP...
-        locationManager.delegate = self
-        //locationManager.requestWhenInUseAuthorization()
-        locationManager.requestAlwaysAuthorization()
-        locationManager.desiredAccuracy = kCLLocationAccuracyBestForNavigation
-        locationManager.startUpdatingLocation()
-        // MAPKIT SETUP...
-        self.myMapView.delegate = self
-        myMapView.showsUserLocation = true
-        myMapView.isRotateEnabled = true
-        myMapView.showsPointsOfInterest = false
-        myMapView.showsBuildings = false
         speechSynthesizer.delegate = self
         uniqueVoice = AVSpeechSynthesisVoice(identifier: daniel)!
         
         // HIDE THE TABLEVIEW INITIALLY
         hideTable()
-        
-        //myMapView.zoomToUserLocation()
-        
-        // LOAD THE ANNOTATIONS!!!!!!
-        // zoomToAugusta()
-        // zoomToMaitland()
         zoomToOrlando(at: 4.0)
+        
+        // MARK: For testing different loactions..
+        /*
+         myMapView.zoomToUserLocation()
+         LOAD THE ANNOTATIONS!!!!!!
+         zoomToAugusta()
+         zoomToMaitland()
+         */
+    }
+    
+    func setupLocationManager() {
+        locationManager.delegate = self
+        locationManager.requestWhenInUseAuthorization()
+        locationManager.desiredAccuracy = kCLLocationAccuracyBestForNavigation
+        locationManager.startUpdatingLocation()
+    }
+    
+    func setupMapViewProperties() {
+        self.myMapView.delegate = self
+        myMapView.showsUserLocation = true
+        myMapView.isRotateEnabled = true
+        myMapView.showsPointsOfInterest = false
+        myMapView.showsBuildings = false
     }
     
     func setUpDisplay() {
@@ -139,7 +134,6 @@ class ViewController: UIViewController {
                     DispatchQueue.main.async {
                         self.setupInitialPoints()
                     }
-                    
                     DispatchQueue.main.async {
                         self.myCollectionView.reloadData()
                     }
@@ -218,20 +212,6 @@ class ViewController: UIViewController {
         }
     }
     
-    // OLD!! ---NOT USING--> MODAL PRESENTATION FOR THE SPONSORED COMPANY...
-    @IBAction func moreTapped(_ sender: UIButton) {
-        guard let cell = sender.superview?.superview as? MyCollectionCell else {
-            return // or fatalError() or whatever
-        }
-        
-        guard let indexPath = myCollectionView.indexPath(for: cell) else { return }
-        let beacon = custAnnots[indexPath.row]
-        let destVC = storyboard?.instantiateViewController(withIdentifier: "DetailsViewController") as! DetailsViewController
-        destVC.thedetails = beacon.title
-        destVC.anAnnot = beacon
-        self.navigationController?.pushViewController(destVC, animated: true)//   present(destVC, animated: true, completion: nil)
-    }
-    
     // NOW USING PREPARE FOR SEGUE INSTEAD OF MODAL
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if let button = sender as? UIButton {
@@ -255,7 +235,6 @@ class ViewController: UIViewController {
         }
     }
     
-    
     @IBAction func showCategoriesTapped(_ sender: UIBarButtonItem) {
         bringOutCategoriesMenu()
     }
@@ -269,20 +248,16 @@ class ViewController: UIViewController {
             self.view.layoutIfNeeded()
             UIView.animate(withDuration: 0.5, animations: {
                 self.categoryMenuConstraint.constant = 0.0
-                //self.categoryMenuVerticalPoint.constant = -100.0
                 self.view.layoutIfNeeded()
             })
         } else {
             UIView.animate(withDuration: 0.5, animations: {
-                //self.categoryMenuVerticalPoint.constant = -550.0
-                
                 self.categoryMenuConstraint.constant = -550.0
                 self.view.layoutIfNeeded()
             })
         }
         areCategoriesHidden = !areCategoriesHidden
     }
-    
     
     //BRINGS UP THE TABLEVIEW...
     @IBAction func showTapped(_ sender: UIButton) {
@@ -319,7 +294,6 @@ class ViewController: UIViewController {
     }
     
     @IBAction func filterTapped(_ sender: UIButton) {
-        
         let checker = sender.tag
         displaySelectedPoints(with: checker)
     }
@@ -394,7 +368,6 @@ class ViewController: UIViewController {
         }
     }
     
-    
     //TAKE A DICTIONARY AND CREATE A CUSTOM ANNOT
     
     func createACustomAnnot(from dict: [String: AnyObject]) {
@@ -427,7 +400,6 @@ class ViewController: UIViewController {
         }
     }
     
-    
     //FIRST INITIAL CONFIG FOR THE MAP AND ZOOM AND CAMERA APPEARANCE
     func zoomToAugusta() {
         let coordinate = CLLocationCoordinate2DMake(33.473244, -81.967437)
@@ -458,15 +430,8 @@ class ViewController: UIViewController {
     func zoomToOrlando(at speed: TimeInterval) {
         let coordinate = CLLocationCoordinate2DMake(28.540765, -81.384503)
         let camera = MKMapCamera(lookingAtCenter: coordinate, fromDistance: 7500, pitch: 0, heading: 0)
-        
-        UIView.animate(withDuration: speed, animations: {
+        UIView.animate(withDuration: speed) {
             self.myMapView.setCamera(camera, animated: true)
-            
-        }) { (true) in
-            UIView.animate(withDuration: 4.0) {
-                // self.createCustomAnnots()
-                // self.myMapView.showAnnotations(self.custAnnots, animated: true)
-            }
         }
     }
     
@@ -582,123 +547,3 @@ extension ViewController {
 extension ViewController: AVSpeechSynthesizerDelegate {
     // All optional methods
 }
-
-// HARD CODED BEACON/ANNOTS DICTIONARY...
-let annotArry: [[String: AnyObject]] = [
-    ["title": "Craft & Vine" as AnyObject,
-     "imageName": "craftVine.jpg" as AnyObject,
-     "beaconName": "sandwich-2.png" as AnyObject,
-     "locatCoordLat": 28.626671 as AnyObject,
-     "locatCoordLong":  -81.435410 as AnyObject,
-     "category": "Food" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2014/06/Craft-and-Vine2-660x390.jpg" as AnyObject],
-    ["title": "Humanitree House" as AnyObject,
-     "imageName": "humaniTree.jpg" as AnyObject,
-     "beaconName": "citysquare.png" as AnyObject,
-     "locatCoordLat": 28.632066 as AnyObject,
-     "locatCoordLong":  -81.431256 as AnyObject,
-     "category": "Historic" as AnyObject,
-     "web": "http://augustalocallygrown.org/wp-content/uploads/xHumanitree_1-1024x768.jpg.pagespeed.ic.UmRq1dFKDX.jpg" as AnyObject],
-    ["title": "Metro Coffee House" as AnyObject,
-     "imageName": "metro.jpg" as AnyObject,
-     "beaconName": "lighthouse-2.png" as AnyObject,
-     "locatCoordLat": 28.635230 as AnyObject,
-     "locatCoordLong": -81.436105 as AnyObject,
-     "category": "Utility" as AnyObject,
-     "web": "http://scontent-atl3-1.xx.fbcdn.net/v/t1.0-9/10891433_10155074380045414_8644919409126086956_n.jpg?oh=aa8dc9887385fb7594335f6741af935a&oe=5A026827" as AnyObject],
-    ["title": "MOD Ink" as AnyObject,
-     "imageName": "modInk.jpg" as AnyObject,
-     "beaconName": "treasure_chest.png" as AnyObject,
-     "locatCoordLat": 28.638891 as AnyObject,
-     "locatCoordLong":  -81.440309 as AnyObject,
-     "category": "Business" as AnyObject,
-     "web": "http://static1.squarespace.com/static/58a8d93ecd0f68a08214dd2a/58a90f99e58c62fa406bfd1c/58aa7228579fb32b000236fa/1487902510435/MODFRONT.jpg?format=1500w" as AnyObject],
-    ["title": "The New Moon Cafe" as AnyObject,
-     "imageName": "newMoon.jpg" as AnyObject,
-     "beaconName": "sandwich-2.png" as AnyObject,
-     "locatCoordLat": 28.5704 as AnyObject,
-     "locatCoordLong": -81.3440 as AnyObject,
-     "category": "Food" as AnyObject,
-     "web": "http://newmoondowntown.homestead.com/files/QuickSiteImages/Menus_01.gif" as AnyObject],
-    ["title": "Mellow Mushroom Pizza" as AnyObject,
-     "imageName": "mellowMushroom.jpg" as AnyObject,
-     "beaconName": "citysquare.png" as AnyObject,
-     "locatCoordLat": 28.5563 as AnyObject,
-     "locatCoordLong": -81.3545 as AnyObject,
-     "category": "Historic" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2014/06/Mellow_Mushroom1-1-DT.jpg" as AnyObject],
-    ["title": "The Pizza Joint" as AnyObject,
-     "imageName": "pizzaJoint.jpg" as AnyObject,
-     "beaconName": "lighthouse-2.png" as AnyObject,
-     "locatCoordLat": 28.5412 as AnyObject,
-     "locatCoordLong": -81.3853 as AnyObject,
-     "category": "Utility" as AnyObject,
-     "web": "http://www.myherocard.com/wp-content/uploads/2016/07/Slider-1-913x240.jpg" as AnyObject],
-    ["title": "Stillwater Tap Room" as AnyObject,
-     "imageName": "stillWater.jpg" as AnyObject,
-     "beaconName": "treasure_chest.png" as AnyObject,
-     "locatCoordLat": 28.5455 as AnyObject,
-     "locatCoordLong": -81.3756 as AnyObject,
-     "category": "Business" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2016/03/stillwater-1.jpg" as AnyObject]
-]
-
-let annotArry2: [[String: AnyObject]] = [
-    ["title": "Craft & Vine" as AnyObject,
-     "imageName": "craftVine.jpg" as AnyObject,
-     "beaconName": "sandwich-2.png" as AnyObject,
-     "locatCoordLat": 33.474223 as AnyObject,
-     "locatCoordLong": -81.966396 as AnyObject,
-     "category": "Food" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2014/06/Craft-and-Vine2-660x390.jpg" as AnyObject],
-    ["title": "Humanitree House" as AnyObject,
-     "imageName": "humaniTree.jpg" as AnyObject,
-     "beaconName": "citysquare.png" as AnyObject,
-     "locatCoordLat": 33.475723 as AnyObject,
-     "locatCoordLong": -81.967896 as AnyObject,
-     "category": "Historic" as AnyObject,
-     "web": "http://augustalocallygrown.org/wp-content/uploads/xHumanitree_1-1024x768.jpg.pagespeed.ic.UmRq1dFKDX.jpg" as AnyObject],
-    ["title": "Metro Coffee House" as AnyObject,
-     "imageName": "metro.jpg" as AnyObject,
-     "beaconName": "lighthouse-2.png" as AnyObject,
-     "locatCoordLat": 33.476423 as AnyObject,
-     "locatCoordLong": -81.969396 as AnyObject,
-     "category": "Utility" as AnyObject,
-     "web": "http://scontent-atl3-1.xx.fbcdn.net/v/t1.0-9/10891433_10155074380045414_8644919409126086956_n.jpg?oh=aa8dc9887385fb7594335f6741af935a&oe=5A026827" as AnyObject],
-    ["title": "MOD Ink" as AnyObject,
-     "imageName": "modInk.jpg" as AnyObject,
-     "beaconName": "treasure_chest.png" as AnyObject,
-     "locatCoordLat": 33.475723 as AnyObject,
-     "locatCoordLong": -81.965796 as AnyObject,
-     "category": "Business" as AnyObject,
-     "web": "http://static1.squarespace.com/static/58a8d93ecd0f68a08214dd2a/58a90f99e58c62fa406bfd1c/58aa7228579fb32b000236fa/1487902510435/MODFRONT.jpg?format=1500w" as AnyObject],
-    ["title": "The New Moon Cafe" as AnyObject,
-     "imageName": "newMoon.jpg" as AnyObject,
-     "beaconName": "sandwich-2.png" as AnyObject,
-     "locatCoordLat": 33.474 as AnyObject,
-     "locatCoordLong": -81.971 as AnyObject,
-     "category": "Food" as AnyObject,
-     "web": "http://newmoondowntown.homestead.com/files/QuickSiteImages/Menus_01.gif" as AnyObject],
-    ["title": "Mellow Mushroom Pizza" as AnyObject,
-     "imageName": "mellowMushroom.jpg" as AnyObject,
-     "beaconName": "citysquare.png" as AnyObject,
-     "locatCoordLat": 33.477 as AnyObject,
-     "locatCoordLong": -81.964 as AnyObject,
-     "category": "Historic" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2014/06/Mellow_Mushroom1-1-DT.jpg" as AnyObject],
-    ["title": "The Pizza Joint" as AnyObject,
-     "imageName": "pizzaJoint.jpg" as AnyObject,
-     "beaconName": "lighthouse-2.png" as AnyObject,
-     "locatCoordLat": 33.473 as AnyObject,
-     "locatCoordLong": -81.966 as AnyObject,
-     "category": "Utility" as AnyObject,
-     "web": "http://www.myherocard.com/wp-content/uploads/2016/07/Slider-1-913x240.jpg" as AnyObject],
-    ["title": "Stillwater Tap Room" as AnyObject,
-     "imageName": "stillWater.jpg" as AnyObject,
-     "beaconName": "treasure_chest.png" as AnyObject,
-     "locatCoordLat": 33.472 as AnyObject,
-     "locatCoordLong": -81.968 as AnyObject,
-     "category": "Business" as AnyObject,
-     "web": "http://2kdda41a533r27gnow20hp6whvn.wpengine.netdna-cdn.com/wp-content/uploads/2016/03/stillwater-1.jpg" as AnyObject]
-]
-
